refactor: content 관련 DTO 패키지 구조 변경 및 Mapper 통합
