+++ android/app/src/main/java/com/on/turip/data/content/dto/ContentPlacesResponse.kt
@@ -1,4 +1,4 @@
-package com.on.turip.data.content.place.dto
+package com.on.turip.data.content.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
