+++ android/app/src/main/java/com/on/turip/data/content/service/ContentService.kt
@@ -2,9 +2,9 @@ package com.on.turip.data.content.service
 
 import com.on.turip.data.content.dto.ContentDetailResponse
 import com.on.turip.data.content.dto.ContentInformationCountResponse
+import com.on.turip.data.content.dto.ContentPlacesResponse
 import com.on.turip.data.content.dto.ContentsInformationResponse
 import com.on.turip.data.content.dto.UsersLikeContentsResponse
-import com.on.turip.data.content.place.dto.ContentPlacesResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Path
